algorithms/src/main/java/top/auzqy/_0001_0050/_0021/Solution_0021.java
package top.auzqy._0001_0050._0021;

/**
 * description:
 *
 *  https://leetcode-cn.com/problems/merge-two-sorted-lists/
 *  21. 合并两个有序链表
 *将两个升序链表合并为一个新的升序链表并返回。新链表是通过拼接给定的两个链表的所有节点组成的。 
 *
 * 示例：
 *
 * 输入：1->2->4, 1->3->4
 * 输出：1->1->2->3->4->4
 *
 * createTime: 2020/4/17 17:33
 * @author au
 */
public class Solution_0021 {

    /**
     * Definition for singly-linked list.
     * public class ListNode {
     *     int val;
     *     ListNode next;
     *     ListNode(int x) { val = x; }
     * }
     */
    public ListNode mergeTwoLists(ListNode l1, ListNode l2) {
        ListNode hummyHead = new ListNode(0);
        if (l1 == null) {
            return l2;
        }
        if (l2 == null) {
            return l1;
        }

        ListNode curr = hummyHead;
        while (l1 != null && l2 != null) {
            if (l1.val <= l2.val) {
                curr.next = new ListNode(l1.val);
                l1 = l1.next;
            } else {
                curr.next = new ListNode(l2.val);
                l2 = l2.next;
            }
            curr = curr.next;
        }

        if (l1 == null) {
            curr.next = l2;
        }else {
            curr.next = l1;
        }

        return hummyHead.next;
    }

}
